Always include 'enabled' field in service response
NOTE: this is basically a mirror of the following patch, but for
services instead of endpoints:

    https://review.openstack.org/#/c/75727/

The 'enabled' field wasn't always returned in a service response.

The 'enabled' attribute for services was stored in the 'extra'
column as part of a JSON string. Now the 'enabled' attribute is in
its own column. It will also be easier to filter out disabled
services now that enabled is a separate column.

With the 'enabled' field being a Boolean column in the database,
the server also needs to validate that the value is a Boolean (if
it's present). This is done in the same way that the server checks
the type of the 'enabled' value when creating users.

Change-Id: I73f0bb58cd31d80b3c1b6c6834e9e8b38adc86a0
Co-Authored-By: Brant Knudson <bknudson@us.ibm.com>
Partial-Bug: 1273867