Cleanup ldap tests (mox and reset values)
When running concurrent test workers, the structure used to test
the instantiation of the FakeLdap handler was relying on mox. In
some cases the test structures were not reset during teardown.

The test cases were not clearing out the _HANDLERS in the common
ldap module. This has been corrected with an addCleanup and moving
the setUp logic to the BaseLDAPIdentity class instead of the
LDAPIdentity class.

Change-Id: I2976774dabeb2b9196b3f25c02838bf4c4688d94