Adding one more check on project_id
Project_id is not a required attribute in creating Trust.
Authentication with this kind of trust results in an error:
"Expecting to find id or name in project".
This message is misleading as the trust creation without project_id
was successful which is a valid behavior.
This error message is generated from checking if project_id is in trust
metadata.
One more check is needed to verify if project_id is None or populated.

Closes-Bug: #1301622
Change-Id: Ic711a63f43ca87a97cb7f761170fa8aa2d45e11e