Fix response for missing attributes in trust
Related-Bug:#1278738

Currently, when the "impersonation" attribute is missing from the trust,
the response shows the error message provided by the backend, which is a
problem since it exposes the backend that is being used. This patch sets
the "impersonation" attribute as mandatory, as specified in the API, in
the "OpenStack Identity API v3 OS-TRUST Extension" document. Thus,
sending an appropriate error response if this attribute is missing.

Additionally, the "trustee_user_id" is now checked in the same way and
keystone will return similar response if this attribute is missing.

Change-Id: Id20fec6798e54d169662b68413591963481405c8