Code which gets and deletes elements of tree was moved to one method
Code which uses 'search_s' and 'delete_s' in ldap backends was moved
to a new method to remove duplication of code.

Added handling of exception ldap.NO_SUCH_OBJECT when we couldn't
find a search base while deleting. Now this exception is
not ignored. Instead of this raised NotFound exception
or more suitable (RoleNotFound, etc).

Also added handling of situation when we trying to delete
nonexisting entries.

Closes-Bug: #1316128
Change-Id: I8e7402337ab4d75ce8920e8158a6e1dd1ed97abd