No longer allow listing users by email
As specified in the bug description, the listing of users by using the
email attribute as a filter would cause an exception. This is because
the email is not a first class attribute, yet it was considered so in
the filters. Thus, this fix will prevent the exception while also
preventing the filtering by email.

Change-Id: Ibf4b51a91c1e47e8197af49855307460559a7fc9
Closes-Bug: #1306835
(cherry picked from commit 75cca803739a5b4ef197666589582802c7391e17)