Refactor test_auth trust related tests
The tests in the AuthWithTrust fixture had many dependencies on global
variables that were not necessarily used; for instance, every time the
setUp function is called, there was a new trust created. This was ran
even for classes that don't need the creation of this initial valid
trust.

This commit gets rid of this behaviour, thus making the test less
dependent on the fixture's state (class variables) in favor of a more
on-demand approach to getting trusts and tokens.

Change-Id: I64752702cc8a328463df15591e3f4c082266ef6d