Use mail for the default LDAP email attribute name
test_user_mixed_case_attribute passes when not using a live LDAP backend.
The test fails when using a live LDAP backend because the standard LDAP
email attribute is 'mail', but the test is expecting the LDAP backend to
use 'email' as the attribute name.  The fix is to use 'mail' as the
attribute returned by the LDAP backend, and 'email' as the corresponding
property in the User object.
To test, run the test_ldap_livetest test e.g.

  python -m testtools.run keystone.tests.test_ldap_livetest

Closes-Bug: #1321822

Change-Id: I2100b5706852fdc7dfea2d4473ef6685c8a2f874