LDAP fix for get_roles_for_user_and_project user=group ID
When there was a role assigned to a group with the same ID as a
user, the LDAP assignment backend would incorrectly return the
assignment to the group when requesting roles for the user via
the get_roles_for_user_and_project method.

With this change, assignments to a group with the same ID are not
returned for the user when calling get_roles_for_user_and_project.

Functions were added to compare DNs more accurately based on the
LDAP RFCs.

The fakeldap code was changed to normalize the values when
comparing values for checking if the values match the filter.

Co-Authored By: Nathan Kinder <nkinder@redhat.com>
Co-Authored By: Adam Young <ayoung@redhat.com>

Change-Id: Ia6f1ae2e3af1e968f1a393bd4f2f38812a88a5d0
Closes-Bug: #1309228