Only emit disable notifications for project/domain on disable
Now that the original (pre-update) version of the project or domain
is available, before a 'disable' notification is emitted, check to
see if the project or domain is already disabled. There is no reason
to perform the 'disable' actions and emit a 'disable' notification
unless the project or domain is changing state from 'enabled'.

Change-Id: I4a7b6a5c638f9faf6bd525553c81e379be562471