Configurable python-keystoneclient repo
The URL for the python-keystoneclient repo couldn't be configured.
With this change, the user can set the OPENSTACK_REPO environment
variable to set the URL for where the python-keystoneclient repo
should be cloned from. So for example you can use a local clone.
The default OPENSTACK_REPO is https://git.openstack.org/openstack
which gets python-keystoneclient.git added ot it to make the clone
come from
https://git.openstack.org/openstack/python-keystoneclient.git.

Change-Id: I936713af3c11e8d68d8739dc68eefabd645df294