Implicitly ignore attributes that are mapped to None in LDAP
Attributes that are mapped to None in LDAP trigger a 500 error when
performing a search if they are not explicitly ignored in keystone's
configuration.  These attributes should always be ignored, even if
the admin left the attribute out of the ignore list.

Change-Id: Ibbabdd0013059d5720250816764021a0b3ce8ce0
Closes-bug: #1335437
(cherry picked from commit 10a3edb7800d380ed72fcbbaae98251998c0e2c7)