Move token_api.unique_id to token_provider_api
Move the unique_id call from token_api to token_provider_api and
mark token_api.unique_id as deprecated, slated for removal in
the K cycle.

As part of the move to non-persistent tokens, the token_api calls
are being moved (as appropriate) to the token_provider_api in
preperation of token_api being deprecated.

Change-Id: I7186b8af3ecc572ca364d58e2ba1a0303e48aca1
bp: non-persistent-tokens