Add the new Keystone TokenModel
This commit adds the new Keystone TokenModel that will be used to
unify in-memory references for tokens. The initial implementation
of this model works similarly to the AccessInfo model from
keystoneclient, providing a unified set of properties that will
return the data from the token-version specific locations.

Future development will round this model out to be better suited
as an authoratative token data source that can be transformed into
any token format / version.

Change-Id: I341e03ac8483f8d6085f4c2c218a28647b9ff2e4
bp: non-persistent-tokens