Expose token revocation list via token_provider_api
Use the token_provider_api for the token revocation list instead
of directly accessing the token persistence service.

Tests have been updated to compare results from the token_api
and the token_provider_api for the revoked token list.

This change is to ensure interactions with the token persitence
are consistent so that it is possible to support toggling
persistence of tokens.

Change-Id: Ic3982329dd28db5b2f974fadbab6c5e010365d27
bp: non-persistent-tokens