Fix revoking a scoped token from an unscoped token
When a scoped token that was created from an unscoped token was
revoked, the original token wound up being revoked. This is because
the scope wasn't included in the revocation event. By including the
scope in the revocation event only the scoped token is revoked.

Change-Id: I5652663ab7e1176d3b1efc5d218a8a020498067e
Closes-Bug: #1347318
(cherry picked from commit da00da732d6be0626810fe4ec87d40dab16117de)