Enable filtering of services by name
This commit enables the filtering of services through the service's
name. This wasn't available before, since the name is not saved in the
backend as a column, but instead it's stored as an extra attribute.

Thus, this commits enables this type of search by asserting a TODO that
was left in the SQL backend, and also adding the `name` attribute to the
filterprotected decorator.

This is needed since some clients (namely python-openstackclient) assume
that this functionality already exists and thus give the option to
create endpoints with the service name as a parameter.

Change-Id: Ifc9c1e44e908410740f6f20380b264d35096fcac
Closes-Bug: #1350273