Enhance V3 version controller to provide JSON Home response
The V3 version controller would only respond with the normal version
JSON response. With this change, if the Accept header is
`application/json-home` it will respond with a JSON Home response
which includes all the registered JSON Home data. Note that there is
no JSON Home data yet so the response is empty.

bp json-home

Change-Id: Icfde89040b376bd6704d497966bd8ad109e32d08