Don't override tox envdir for pep8
infra's run-pep8 job assumes the envdir for pep8 job is not
overwritten, so it can call out to .tox/pep8/bin/pip.

This bug is causing keystone pep8 job to occasionally fail in the gate.

If the motivation behind overriding the envdir jobs to use venv, one can
use devpi and/or wheels to speed up the process of building a venv.

This isn't a backport because in stable/icehouse Keystone didn't set
the venv for the cover job.

Change-Id: I32cd46152caa73e1308fdd216ba016c80a5a9b4f
Closes-Bug: #1352134