Remove identity_api dependency on token_api
Remove the identity_api dependency on token_api. Now for each type
of event that would cause revocations (or delete of tokens) to occur
an internal callback notification is emitted and consumed by the
token_provider_api and the revoke_api (as appropriate). This required
a few new notification entity types ('user_password',
'user_removed_from_group'), but since these are not public notifications
the overhead should be minimal.

This change is to ensure interactions with the token persistence system
are consistent and supporting the ability to toggle token persistence
in the future.

Change-Id: I796f1b003a9d43b89940bb74d2c44d3cff215c72
bp: non-persistent-tokens