Remove wsgi and base controller dependency on token_api
The wsgi module and the base controller no longer depends on the
token_api, to interact with the token persistence system. The
token_provider_api.validate_token method and the KeystoneToken model
are now used.

This change is to ensure interactions with the token persistence system
are consistent and supporting the ability to toggle token persistence
in the future.

Change-Id: I5831d6a086241bc3a6691602e7527ae53c9411ce
bp: non-persistent-tokens