Notification Constant Cleanup and internal notify type
Added a new notification type 'internal' that will never be public
which is utilized for communicating certain things exclusively to the
notification callback system where there is no overlap with the
public notifications (to be put on the bus) or if there are special
case notification data types.

Basic cleanup on constants within the notification callbacks has been
done to utilize a centrally defined set of notification types and
internal-specific user token invalidation. Forward looking some more
constant cleanup will need to occur. Eventually there will be
additional internal-notification types.

This is all in support of removing the dependency of moving to
non-persistent tokens which requires unifying token persistence
management.

Change-Id: I16b608e8502f38c3a88579aa1bc91b4af7ed7bc9
bp: non-persistent-tokens