Enhance GET /v3 to handle Accept header
Keystone wasn't handling the Accept header for GET /v3 according to
the HTTP 1.1 spec[1]. This change enhances the Keystone server so
that it supports using multiple values and qvalues in the Accept
header for GET /v3. This allows requesting a JSON Home or JSON
response with a preference for JSON Home, for example.

[1] http://tools.ietf.org/html/rfc2616#section-14.1

bp json-home

Change-Id: Ifb204b6860c61a874ecbf92388878685ac437987