Generate IdP Metadata with keystone-manage.
For proper federation setup, an Identity Provider peer needs to
generate metadata that can later be consumed by a federated Service
Provider. Since metadata content doesn't change very often it's
advised to generate a static file and later read its content on demand
(this is out of the scope of this patch). This patch adds required
functionality for utilizing the federation.idp.MetadataGenerator class
from a keystone-manage command line tool. Generated metadata will
be redirected to the stdout.

Change-Id: Ifed1f4e4ede0276b3aba22aa9e236e5e74e2ed93
Implements: bp keystone-to-keystone-federation