Remove SAML2 plugin dependency on token_api
Remove the SAML2 auth plugin's dependency on the token_api by utilizing
the token_provider_api.validate_token instead of token_api.get_token and
leveraging the KeystoneToken model instead of doing direct lookups on the
dictionary returned.

This change is to ensure interactions with the token persistence system
are consistent and supporting the ability to toggle token persistence
in the future.

Change-Id: I1a38156ff5535e1dd52460040891ad50500d4c71
bp: non-persistent-tokens