region.description is optional and can be null
region.description is documented to be optional and should be treated as
optional.

A null description provided when creating indicates that the user
doesn't want a description stored. This is equivalent to not sending
a description property at all.

A null description provided when updating indicates that the user
doesn't want a description stored and any existing description should be
removed.

In both cases we treat this as an empty string because our database
schema requires a non-null value for description.

Closes-Bug: #1398165
Change-Id: I4a890414927ab5278861341b71a3e7fb324946a4