use one indentation style
There are two (or three, depending on how you count) indentations styles
mixed together in this one logical line of code, which are further mixed
with two methods of closing parens and curly braces.

Besides being difficult to read, the mixture resulted in an off-by-one
error that isn't caught by the version of hacking that Jenkins is using,
but is caught by the latest.

Change-Id: Idab3c80c720e0bfd5741dc9e07f841b5ac270960
Closes-Bug: 1363019