Fixes a mock cleanup issue caused by oslotest
The latest olsotest includes a call the mock.patch.stopall. If a patch's
stop is registered as a cleanup before oslotest add its cleanup the
following traceback will happen:

    local/lib/python2.7/site-packages/mock.py", line 1376, in __exit__
        raise RuntimeError('stop called on unstarted patcher')
    RuntimeError: stop called on unstarted patcher

This is because the cleanups are handled by testtools in reverse and by
the time our cleanup runs the patch would have already been cleaned up.

Change-Id: I6df0a267420eec794582acdf0429a141e7f64e45
Partial-Bug: #1365678