Fix using local ID to clean up user/group assignments
When using ID mapping and a user or group is deleted, the mapped id
was being passed to the assignment backend when it's told to clean up
the role assignments for hte user or group. The role assignments are
based on the external ID and not the mapped ID, so the role
assignments wouldn't be cleaned up.

With this change, the external ID is passed to the assignment backend
methods.

Change-Id: I4405329df509f81c9c1178e91092955b16cd586b
Closes-Bug: #1222675