Ensure a consistent transactional context is used
This change repairs a migration that was using multiple
transactional contexts for a single operation.  Using
a Session implies that operations do not proceed under
"autocommit", so here the autoload operations as well as
migration operations are invoked under a single
connection.  The issue has been exposed due to improvements
in oslo.db which expose SQLite as supporting transactional
DDL.

Change-Id: Ic6673234122e1cf1bc4c05907299bb1938e54a9f
Related-Bug: #1367354