Add info about pysaml2 into federation docs.
pysaml2 is only required for Keystone2Keystone federation. Due to code
structure, ``pysaml2`` needs to be installed in order to enable
federation extension. Since federation is an extension all the
dependencies are not in the requirements.txt file and will not be
installed by default. This patch updates federation docs adding
information that the library must be installed by hand.

Change-Id: Icc0b3b5cc0e22950a85bb58bbd3631c911be8621