Fix tests comparing tokens
There were tests that verified that the PKI token body could be
encrypted with CMS and compared to the token ID in the response.
This test isn't safe because the token body may be different than
the token encrypted with CMS since the order of items in the dict
can change.

The fix is to change the test to decode the PKI token ID and
compare that to the response body JSON instead.

Change-Id: Icc649b96071ff084d5c76f2ea2bcf3ecb08a0351
(cherry picked from commit 83394ba71ba06e9899aaa62f211919075f2a3d91)