Adds a wip decorator for tests
This decorator can be used to commit failing tests while they are still
in development. It can also be used by people to show how an issue can
be reproduced without them having to make the code changes necessary to
make the test pass.

This is nicer than just raising TestSkipped because there is a built in
reminder to remove the decorator when the test starts passing.

Implements: blueprint failing-tests
Change-Id: I9ded266b368e7955b1e295950df394823b1a4088