Sync modules from oslo-incubator
Sync the following modules with oslo-incubator commit
2acb6502268f7cc973c2b37ee49b73de81b8b69d :

- eventlet_backdoor
- fileutils
- log
- loopingcall
- policy
- processutils
- service
- versionutils

Use mask_password from strutils.

Remove excutils, jsonutils, strutils, and timeutils since those
modules aren't used anymore.

These modules are chosen because they're easy to sync. The goal of
the next sync is to remove the rest of the deprecated modules, but
this is blocked until Keystone uses the oslo.config config generator.

Change-Id: I7a03842d800a94a7a231da2a497be8937318857d