Move eventlet server options to a config section
Keystone can run in an eventlet server (keystone-all), or it can
run in a WSGI container (Apache Httpd). There are several
configuration options that are only used to configure the eventlet
server and are ignored when running in a WSGI container. Having
all the eventlet server options in the default section makes it
difficult for deployers to know what can be ignored when running
in WSGI mode. The options that are only used by the eventlet
server are moved into an [eventlet_server] section, and the SSL
options are moved into [eventlet_server_ssl].

bp eventlet-server-config

Change-Id: I6dd718c4d54056d0e29978f393ec45f7291f802d