Prevent infinite loop in token_flush
If there were > batch_size tokens with the same expiration time then
token_flush could hang because it picks the expiration time as the
target to delete with a < check.

The check is changed to a <= so that the tokens in the batch are
deleted so that token_flush always makes progress.

Change-Id: Id0b6f7fcdf7ebb02a6f38fbb9a4e25bf714a614c
Closes-Bug: #1387401
(cherry picked from commit 6bf78255b2529fd3a83e749101130cf595f1544e)