Use _ definition from keystone.i18n
Currently there is one case where the _ builtin is being used
instead of the translation function defined in keystone.i18n.

This will cause the messages to not support lazy translation.

This was not being checked by pep8 because the _ was specified
as a builtin in tox.ini's flake8 section.

This patch updates tox.ini to remove the builtin specification in
the flake 8 section to enable this check.

'from keystone.i18n import _' was added to the file that was
identified by the updated pep8 checking.

Change-Id: If30e394bde4b75fa1067e910a1e73a95e8e6c5dc